+++ kwcoco/cli/coco_eval.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python
 """
 Wraps the logic in kwcoco/coco_evaluator.py with a command line script
 """